Makes service calls more robust in error conditions
If a service fails, an automatic error is created and sent. However, if the service already created
enough data (8k) the response is already committed and no error message must be sent. Rather
we have to ensure that the underlying request (its OutputStream) is closed properly and a warning
is issued.