Adds a new HttpDataFactory which deals with empty fields properly
When receiving a POST, its name value pairs are decoded and
represented by Attributes. Those attributes are created by the
HttpDataFactory. Now gifted clients (like ICEFaces) create
invalid POST data which empty field names (&=...). This
is prohibited by the default implementation and the request fails.

We fix this by replacing the empty name with "_sirius_unknown" which
is hopefully not used elsewhere.