Changes the source of the default config for scopes.
Previously the default config was nested in the application config.
But since there will be a lot config values, we now load the scope
default from files in scope/*.conf. Also each customization
can provide a scope-settings.conf. Furthermore a GUI was
created to view the default files as a source of inspiration when
provides a custom scope config.