Makes the server more HTTP compliant by sending a Connection: Close header
If we do not support any more keep alives for the current connection,
we will close it (as specified in the standard). We now also indicate this by
adding an appropriate (Connection: close) header for HTTP 1.1 requests.
HTTP 1.0 had no keep alive and would only be confused by this behaviour.