Fixes a threading issue which might lead to very strange behaviour as the CallContext vanishes while processing a request.
As netty re-uses threads, we have to re-install the CallContext on each invocation of the user code. Otherwise
a request, with a large upload might loose its CallContext as it was only installed in handleRequest, not when
handling the last http content. So an intermediate request handled by the same thread overwrote the request with
an already handled one.

We also added some null checks for good measure - when encountering an inconsistent state, we rather close the
channel an get rid of all state thant trying to compute a meaningful response. This shouldn't happen in the first case.