Fixes support for web sockets.
We need to initialize the handshaker with the effective URI (including the query string) as it now
performs a rather dumb comparison to determine if a handshake should be attempted.
As in our case, the whole protocol handler is only injected into the pipeline if necessary,
we don't need this check and can therefore bypass it by using the uri itself.

Also we need to modify the HttpPipeliningHandler to forward messages which are received after
the request was completed (LastHttpContent). We do that now as long as no pipelining conflict
occurs.