Fixes a nasty race condition which caused a NPE when an idle timeout occurs.
As we first released the post decoder and set it to null afterwards,
another thread could still obtain a reference to the now inconsistent post decoder
and then fail miserably. Although this exception was catched and handled properly
(the attribute returned null) it still left nasty stacktraces in the logs.