User agent detection
From now on it is possible to determine if the current request was created by a deskop client or a mobile client. It is possible to distinguish between phones and tablets for mobile requests. Furthermore it is also possible to determine if a request was made by an Android device or an iOS device.