If an inlined template has local variables, we have to shift them on our stack.
For each template being inlined, we need to allocate a stack location on the caller
stack and we also must update all references to it. As processing the arguments
might also create temporary stack locations, we first transfer the locals (reserve
room for them) and then port the arguments.

The order doesn't matter technically, but might look confusing. However, this is the
only simple way to ensure that no conflict / mis-numbering occurs due to a newly
created temporary being moved "twice".