Fixes a crash due to premature access on a post decoder.
If we log an error during the processing of a request, i.e. an SSL
problem. A LogTap (like on in sirius-biz) might try to identify the
associated user. Thus it tries to access the POST decoder, which is
not completed yet and throws an appropriate exception. We ignore
this exception, as it only jams the logs. Sadly, there is no way,
to ask the POST decoder if it is done yet or not, which would be
a bi cleaner.