Prevents a potential stackoverflow when finding the current user or scope.
If an error happens while binding the current request to determine the
current user or scope, an error handler might call getUser and getScope
again to create an appropriate log message (sirius-biz does that).

If case of an recursive call, we simply provide a fallback value
(UserInfo.NOBODY / ScopeInfo.DEFAULT_SCOPE) to prevent further recursion.