Shorten uri written to TaskContext as job
The TraceData in Sirius-Biz saves the origin of a change. This contains the system string from the corresponding task context. As some requests might contain long parameters, this would also be saved. However, as the field has only a size of 150, longer calls with many and long request parameters result in an exception. As the requesrt uri also might contain sensitive data, this is now stripped by only setting the uri without the request parameters in the TaskContext in the WebServerHandler like the ControllerDispatcher already does.