Cleans up the caching used by AssetsDispatcher.
Currently invoking the dispatcher without any settings, should have applied a no-cache header.
This failed whenever we delivered a file (scss / static file) as this applied a cache setting by itself.

Also it turned out to be quite a bad decision not to cache assets at all.

We now apply the following policy:
- plain URLs (/assets/....) are cached for 1h by default
- dynamic urls (/assets/dynamic/TOKEN/...) are cached FOREVER
  as the token will chance with the next patch / release
- /assets/no-cache/... will not be cached at all.