Recompute CSRF tokens in advance
Currently the CSRF tokens can expire by staying on a page for a few minutes in the time window they expire. A request in this scenario would submit an expired CSRF token which can cause errors. Therefore CSRF tokens will now be recomputed in advance to avoid them expiring while staying a few minutes on a site.

Tags: csrf