Improves the performance measurements done by the webserver.
Previously the webserver measured the time to first byte (TTFB)
which is the time, the webserver took to generate the first byte
of the resonse - which is equivalent to the actual execution time
of a request without measuring inbount or outbound bandwidth.

However, if the webserver is overloaded, requests get queued
and this time will increase - thus the server will report false
positives as long running requests.

This is solved by performing two measurements. Once as the
requests ist submitted and another one when the actual execution
starts. We can now report the queue time (wait time) and measure
the correct TTFB - thus identify those reuqests slowing down
the web server.

We also report all parameters of a long running request to
further support debugging.