Code cleanup
Removes a bunch of unused code. As we now (actually for some time)
use a cookie storing the client session, we no longer need an extra cookie
storing a login token. Hence this code was removed.

Also the "trusted" flag died in WebContext, therefore there are no extra
"trustedRoles" which weren't used anyway.