Proper cleanup of locals after for-loop
This commit changes the cleanup of local variables after a for-loop using the for-handler ('@for') or for-tag('<i:for>').
The problem was, that only the last element was popped of the stack, which only works if the loop just introduced the
loop variable but no additional local variables using '<i:local>'.