Proper cleanup of locals after if-else-block
This commit changes the cleanup of local variables after a if-(else)-block using the if-handler ('@if') or if-tag('<i:if>').
The problem was, that no elements are popped of the stack, when leaving such a block, which can lead to template
compile warnings and allows to falsely access a variable out of scope.