Permits to perform "session pinning" with the help of a second cookie.
This will be stored next to the session (as separate cookie) and also within the session. This greatly protects against
session hijacking (or software bugs with the same effect) as both cookie are never set in the same request / response.

Also, an MD5 and not the pin itself is stored in the session, therefore it is not enough to obtain only one cookie.

Also some tracing and logging was added to help troubleshoot session related bugs.