Only updates / sends a set-cookie for the client session if the response isn't cacheable.
Everything else might theoretically leak the session (although a lower
protection layer will mark the request as "no-cache" anyway). However,
this will yield a warning in the logs. Since, due some side-effect, the
session might be rightfully changed and this change might also be happily
discarded, we only log a developer info and omit the set-cookie entirely.

This will also ensure, that no session-pinning is attempted in such a
response - which is equally bad and opposite to its purpose.