Remove call to "suggestions.getTokenForValue"
Because it could be that, no suggestion exist for this value and so the method returns undefined. (When we enter a text into the tokenfield which is not in sugestions it still creates a autocomplete row for the text if we allow custom entries. See line 737 of this file)

Also we can simply add a string here, it does not need to be a token