Fix bug, that completion could not be selected by mouse after searching
We do not need to set "completions.selectedRow = undefined;" in the "show"-method because onHoverRow is handling the unselection! In IE the events were a diffrent order, so this messed it up a little bit.

We add tokens on the "onSelect" because then the order of events is more reliable in IE