Improves predictiveness of blur events by defining tabindex for elements
Some browsers wont fill the relatedTarget property of the focusout event if the selected target has no tabindex.
The relatedTarget is usefull to determine what element was clicked to trigger the blur event on the input field.
The autocomplete field for example uses this to prevent the blurring of the tokenfield when the autocompletion box, a suggestion or its scrollbar is clicked.

Fixes: SIRI-45