Prevents token autocreation by tokenfield when using the keyboard
This is done by blocking the keydown event when a autocomplete suggestion is selected by keyboard and enter is pressed.
Tokenfield created a duplicate token when choosing a autocomplete suggestion with the keyboard.
So after entering '123' and selection the '12345' selection the tokens '123' (wrong) and '12345' (correct) were created.
Now only the correct token is created.

Fixes: SIRI-45