Prevents token autocreation when autocomplete box is clicked
Clicking the autocomplete box, a suggestion or its scrollbar resulted in the creation of an unwanted token by the tokenfield.
This is done by toggling a flag when a click on one of these elements results in a blur and not adding the next token when this flag is toggled.

Fixes: SIRI-45