Prohibits keepalive if a response is created before the request (and its content) has been fully read.
If we respond to a request before it was fully read, we have to give up keepalive and should
actually close the connection, otherwise we might confuse HTTP clients like curl.

This behavior was detected when a pre-dispatched controller tried to prevent an upload
by sending an error which was actually encoded as 200 OK + JSON response. Therefore
neither the HTTP client nor any intermediate proxy saw any need to abort the connection
and the upload of the data continued after the response was received. However, this
confused some client like curl or essentially libcurl.

Fixes: SIRI-47