If a response is trying to be created before a request was fully read, we now exhaust (throw away) all remaining data.
HTTP (at least in most implementations) does not support premature HTTP responses before the request
was fully read. At least not when responding with a 200 OK. However, we do not want to send an explicit
HTTP error, as in this case some clients cannot parse the resulting JSON.

Therefore we now discard all incoming data and then send the response.