Also output all failing metrics when reporting an unhealthy cluster.
This is quite useful as some monitoring tools (e.g. UptimeRobot) contain the monitoring URL in their alarm message.
Therefore one can quickly determine which metric is failing.

Note that we output these in lowercase as we don't want any conflicts for the "OK" keyword. We also use 417 as
HTTP status code (5XX would be more appropriate) as we want to report an "non OK" response but we also
do not want an upstream reverse proxy to "retry" another node or anything the like...