Adds the streaming XLSX workbook mode to the excel exporter
The SXSSF workbook is useful when writing large amounts of data which exceeds the amount of available heap. In contrast to the normal XSSF workbook which would eat up all heap and then fail the SXSSF workbook writes the exported rows to disk periodically which drastically relieves the heap usage and allows to export huge data sets.

One caveat is that the streaming mode can in some cases produce excel files that are incompatible with some clients. Because of this the SXSSF implementation is not used by default but the accessor of the exporter API has to make the choice which implementation is to be used.

Fixes: OX-5454