Does not log an invalid CSRF token into the system log.
The intention was to log this as this might be a security
breach / attempt. However, it turned out that in most of the cases,
the user just used multiple tabs or did other wired stuff with their
browser - therefore there is no need to jam the system log with these
messages.

Fixes: SIRI-81