Actively flushes the bytes out into the network if .flush() is invoked.
If a ChunkedOutputStream is created using Response.outputStream
and then flush() is invoked on it, we previously only wrote the buffered
data into the underlying channel buffer and let Netty and the OS decide
what to do with this. We now enforce a flush on the channel buffer as well
so that we make sure that the data is actually sent out into the network.

This might be required if we rely on exact streaming behavior (e.g. when
creating a streaming JSON response).