Installs a circuit breaker.
When trying to invoke bindUserIfPresent(WebContext) while we're
already resolving the current user, we must abort to prevent an
endless recursion.

This happened in SellSite where we tried to obtain a database,
which in its constructor created a formatter, which in turn detected
the current language which then eventually invoked this method
when then again tried to resolve the current user, ending up trying
to obtain the same database again.

As databases are kept as ConcurrentHashMap, re-entering computeIfAbsent()
runs into an endless recursion.