Also detects unnecessary casts when the source is Transformable.
Even when using transformers a call like X.as(X.class) is pointless.
Also this could otherwise cover follow up errors as previously
(before Tagliatelle supported generic propagation), we operated
on Objects which were often casted to Entities of sirius-db. Now
we know it's an entity (which is also Transformable) so the unneccessarry
cast isn't discovered but executed a pointless Transformable.as().