Uses MethodHandles in favor of classic reflection.
This performs all accessibility checks during the startup
and invoking a route is more or less directly translated
into a virtual method call. Therefore we also unfold the
arguments and provide a direct invocation for 0-4 arguments.

This permits to use a more efficient call into the JVM. Only
routes with more than 4 parameters take the burden of
creating a argument array and invoking a less performant
variant, which more or less unfolds the given array again...