BREAKING CHANGE: Uses a simple loop instead of a recursive callback monster.
We previously retrofitted the dispatchers to not only skip
a request and let the next dispatcher given it a try, but also
to "restart" the whole pipeline and start from the beginning
(most probably after the request has bee rewritten).

To make all this way simpler (and more obvious) a special
return type was added in favor of additional callbacks.

This will also significantly reduce the length of stack traces
generated from within the web server.