Adds a hidden input which is always submitted and thus shows the presence of a checkbox.
If a checkbox isn't checked, the client won't send any indication to the server, that the
checkbox was present in the form (or not). This leads to problem in the BizController
when performing a load / autoloading as boolean flags get reset as we don't know if
the box was present in the UI or not.

This is especially a problem, when a complex form is split over several pages.