Provides a keyword handler for invoke and include.
This is especially useful when combining multiple JS files or other resources which
aren't XML based. Instead of writing <i:include... in the middle of a JS block, one can
now use ___invoke or ___include.

The former will include a pasta template and also compile it, the latter will only include
the literal contents of the given asset.