Removes the ability to reference a helper by its name.
This is essentially a deprecated approach which was left over
from the JSF days. We actually never did reference helpers by
their name and we probably will never, as we'd instantly loose
all benefits of Tagliatelle and its static compilation capabilities.

Note that now that HelperFactories are quite trivial as these are
only glorified constructors, we also permit to instantiate a helper
via its constructor (which might take ScopeInfo as parameter).

Note that a factory can still be used and will always be preferred
as we might use this to overwrite / change helpers in customizations.

Also note, that while refactoring the new instance creation, we also
added support for circular references across helpers and also made
sure that each helper only exists once per scope. (Previously the
@Helper annotation created a copy instead of referencing the real
helper).