Replaces MethodHandles by MethodPointers.
We need to use this custom type so that constants are
properly de-duplicated (MethodHandle has no equals).

Also we can provide a proper string representation (which
is also pretty short when using MethodHandles).

Note that this is one of the fixes we used to prevent a Noodle
VM crash, as the constant index overflowed into the op code
value within a bytecode.