Fix the JMP_FALSE handling
The JMP_FALSE opcode pops a value from the stack and will jump to a
given label if the value is false. In the AND operation this will
prevent the evaluation of the right operand and the whole operation will
evaluate as false (See Conjunction.java). Likewise in a OR operation the
JMP_FALSE will prevent the evaluation of the second operand and the
whole operation will evaluate as true in case the first operand is
true (See Disjunction.java).
In the case of 'null AND true' the null is not handled as being false,
leading to a true evaluation of the operation. To fix this an additional
check was added to handle null as false when the JMP_FALSE opcode is
encountered.

Fixes: SIRI-317