Extracts the PDF image replace logic into separate handlers
This allows for easy integration of custom logic (for example embedding
blobs without the use of http).

Migration:
If internal resources (assets) are used via relative URLs these need to
be prepended with resource://. Barcodes and QR codes using the type
attribute in the img tag are now instead generated by using the
barcode:// and qr:// URI protocols.

Fixes: SE-8489