Implement a crude fix
The problem surfaces when we try to resolve the type of a method call.
In the case of Optional<String>.empty().orElse(null) we already know the
return type (T) will be String. But our logic throws this info away and
replaces it with void when propagating the type of the null parameter
value.

Fixes: SIRI-321