Fixes a potentiel memory leak.
If we hand out a ByteBufInputStream, we rely on the caller to guarantee that
the stream is closed under any circumstances. As it turns out, this is kind
of an optimistic approach.

As we know, that the buffered data isn't that large (it would have been moved
to a file otherwise), the create a GC managed copy of the data - just to be on
the safe side...