Support resolving the return types of the whole inheritance tree
In order to be able to resolve the types from methods that are defined
in super classes or interfaces, we need to also add their type
parameters. This is done recursively to support more than one level of
inheritance.

This added the requirement to scope the type parameters in the type
table, as (like in our test) the same letter might be used multiple
times with different types.

Fixes: SIRI-346