Filter out bridge methods.
These methods are synthesized by the javac compiler to handle covariant
return types. Sadly this creates a method list which cannot be created
in a sane environment (e.g. two methods can have the same name and
arguments but different return types). These methods are only generated
to coerce calls to covariant superclass methods.

We're generally not interested in these methods, as these also wipe out
any generic infos.

Note that we cannot provide a test for this fix, as the method list of
a class is unordered (ordering is sheer luck). Therefore most of the
time the bridge methods are placed after the "real" methods and thus
this fix would not be needed. However in *some* circumstances, most
commonly in production, the order seems to be flipped.