BREAKING CHANGE: Refactors UserContext and its Message(s).
We previously more or less expected messages to be in plain text and
to optionally contain a single link.

We now refactored to inner workings of the whole framework to
essentially support any HTML contents as message. We also provide a
builder to properly control and decide if HTML content is present or
if plain text is being output (which requires proper escaping).

This also opens up the possibility to further process some error
messages (e.g. those by Tagliatelle) to ensure proper rendering.