Provides a new way of providing JSON and XML services.
Using the annotations @PublicService or @InternalService, one can define
that a route is considered a service call with a certain response format.

Previously this was done using Routed.jsonCall. We now use extra
annotations to support XML as additional format.

Using @PublicService also provides the benefit of self-documentation
as these services are listed in the Public API explorer.