Improves error and failure handling in the BasicController.
If the response is already committed, we do not need to care about
the HTTP code, we simply have to submit an error so that the underlying
channel is closed.

Also, if the response for a service call was already commited as the
internal buffer is full, we can no longer report and error. This
(edge) case is now logged at least to support debugging.