Installs a custom browser history manager.
This is used to customize and also remove some history entries. As the
"real" history API doesn'T support skipping or removing history entries
(e.g. for routes like "/.../new", we install a custom history stack
stored in the localStore.

If this fails, we resort the the normal history.

Note that we also provide a "t:backButton" tag, which only shows the
back button if there is a app-local URI to go back to. This permits to
show an "intelligent" back button for the 404 page.