Permits to fix and reload broken extensions in dev systems.
We previously forced the list of extensions for a given point to be
empty, if one of the templates failed during compilation. We now
permit the developer to fix the template and we'll then recompile all
extensions properly.

Note that we still keep the empty list around in prod systems, as there
is no way of healing a broken template.