Attempts to handle "/" as "/admin" before resorting to a 404.
Tycho based backends commonly use /admin as the main or at least the
alternative endpoint to show a dashboard etc.

If an application doesn't handle "/" it probably wants to use the
provided dashbaord. Hence we restart the pipeline with the given
uri.

Note that this is done in favor of simply providing a low priority
route in a controller, as an application might handle "/" via a
dispatcher which runs after the ControllerDispatcher.