Removes an unnecessary call to strip.
This is unused as we do not expect the given path to contain a query
string at all. Also we obviously do no expect the path to contain
any encoded characters, as otherwise these had to be decoded into
requestedURI. Therefore we should be able to directly use path as
rawRequestedURI here in a safe manner.