Enhances history handling and make it cross-compatible with wondergem
If you're not using tycho yet, you should disable tycho history in wondergem via config.

The goal of the own history handling:
- Skip over multiple entries of the same page
- Skip over entries that explicitly want to be skipped (e.g. historyUri = '' in t:page)
- Redirect to a different historyUri

As the browser history API is pretty restricted and popstate calling is pretty inconsistent we implement this by reading state on page load and redirecting / skipping over urls as instructed.
This leads to some restrictions:
- Forward Navigation will only work once
- When reloading a page which already has multiple entries, a forward button will be shown, which doesn't do anything
- Navigation using the back menu (more than one step back) might fail in some rare circumstances