Disables autocomplete on hidden select fields
As this will also disable caching the field value on page reload in Firefox. This feature might lead to Firefox selecting the first value on page reload even if no value was selected before the reload. It's not quite clear what's causing this weird behavior, we can just disable the autocomplete as we don't need it anyway on the hidden field.

Fixes: OX-7432