Introduces JS and HTML style comments which are recognized by Tagliatelle.
Normally Tagliatelle ignores all kinds of comments and outputs them
right away. This is mostly useful (e.g. some still comment out <script>
tags).

Now in order to mark comments as true comments to be fully ignored by
Tagliatelle, we introduced to sequences: /**@ will ignore everything up
to the next */ and <!--@ will ignore everything up to -->.

This is especially useful to provide elaborate comments in JS libraries
which are then discarded in the output (in the name of SEO / download
speed etc.).