Properly implements the "default-template" mechanism for extensions.
If a template in "default/extensions/name" and one in "extensions/name"
exists, we now only load the second (in accordance with all other places
where default templates are overwritten by ordinary ones with the same
name).

This way a default extension in a library can be provided which can be
overwritten in a product).