Performs a block-wise trim for GlobalRenderContext.emitToString.
This way, content of blocks is output independent of the indentation
of the source-code.

Therefore:
<x:foo>
   bar
      <div>baz</div>
   bar
</x:foo>

Will return

"bar
   <div>baz</div>
bar"

in renderToString. The same optimization is used when accessing
the contents of i:define or i:pragma.

This is e.g. used in sirius-biz for <kb:code> to properly layout
nested code block with proper indentation.