Entirely kills contention protection when writing into channels.
We introduced this back when we tried to handle some dispatchers
directly in the event loop of Netty. This turned out to be a dangerous
game all together. Therefore, we moved dispatchers into their own
thread-pool. With this new scenario, writing into a channel is
an entirely different beast, as the event loop can shovel away data
while we're writing into it.

ALSO, when tunneling data, all callbacks into the TunnelHandler are
executed directly in the event loop of the HTTP client, which means,
that the contention prevention will effectively halt all tunnels -
with all the horrible effects imaginable. Therefore, blocking as been
removed and we rely on the channel buffers to keep everything working.