Prevents long iconInlineInfo values from breaking out of cards
The iconInlineInfo component is most of the time used inside of cards to show various information about the entity the card is about.
Having long text (like a url or filename) without easy ways for CSS to determine where to wrap the text meant those could easily break out of the card body.
We now tell CSS it can break the value wherever necessary.

Fixes: OX-8531