Further narrows a HTTP header sanity check.
We observed horrible behavior of downstream proxies if both, set-cookie
and Cache-Control/Expires headers (which instruct a proxy to cache the
response) are present. As per HTTP a response containing a set-cookie
header must never be publicly cached, however, some did.

We therefore built a detection to recognize and warn about hazardous
header combination. However, this check was too harsh and warned about
some "safe" scenarios. This commit tries to detect and accept these.