Passes array parameters as multiple single parameters
The default behaviour of FormData when an array is passed as value of a parameter is to transform the entries into a comma seperated list which is sent in a single parameter to the server.

jQuery AJAX did this differently and sirius server also expects is in a different way. We therefore send each entry of the array as a separate parameter (with the same parameter name) in the form to the server.

Fixes: OX-8650