Properly kills the ChunkedWriteHandler out of the pipeline after it has been used.
We want to get rid of this handler, as it violently stores written data
(in its queue) and even drops/ignores flushes if the channel isn't
writable right now.

This is mainly a problem when using the tunnel handler, which doesn't usually
generate a flush event for every write.Therefore, we ensure that the handler
isn't present at all - as we don't need it anyway.