Forces a flush every now and then once we know, that "enough" data has been written.
This is to ensure that any intermediate handler will flush any buffer or the like
frequently. Sadly, this isn't enough the flush the ChunkedWriteHandler, as it
only properly flushes, if the underlying channel is writable.

Therefore we get rid of this handler during tunneling entirely (done in a separate
commit).