Uses initial as max height after opening accordion content
This better represents the state we wan't to reach (the initial max height of the element - no max height constraint at all).

Unset behaves a bit weird, as it it differentiates between inherited and non-inherited properties which makes it kinda hard to predict. In the case of max-height it would result in the same behavior though. So this is just a change for more clarity :)

Fixes: OX-8816