Allows to override the path that is used to generate deep-links
The use-cases of this could be described as edge cases but it is quite necessary for how the OXOMI frontend used the history to store its state and navigation.

This new method can be used to override the path that is included in the generated history state URL externally. For example when sub-navigation occurs that should be reflected in the path.

Fixes: OX-8795