Allows to block closing of overlay via ESC key by providing callback
Introduces a new optional config param that allows to pass a function that gets called when the user presses the ESC key and can determine whether the overlay should be closed at that time or not.
This can be useful when the content that is rendered in the overlay uses the ESC as well.

Fixes: OX-8917