Allows to detect an overlay actively being closed by the user
We trigger the `sci-overlay-destroyed` in the following scenarios:
- Overlay closed via x button
- Overlay closed via background click
- Overlay closed via ESC key
- Overlay closed as new one was opened

The last one of these is the only one that is not triggered by the user actively trying to close the overlay but can also occur when the currently open overlay triggers a new overlay that replaces the old one.

To differentiate these we now trigger a new event `sci-overlay-dismissed` in those 3 cases where the user tries to close the overlay.

Fixes: OX-8917