Throws a proper exception if the underlying channel is closed
NIO and Netty commonly use this exception to signal, that
the underlying channel has been closed. We therefore throw
the same exception, so that our handling code (e.g. in
ContollerDispatcher.java) can detect this and properly ignore
such errors (or at least not write them to the syslog) as this
is an behavioral error, not a system error itself.

(cherry picked from commit dc041e145645d1c1322690e93d0a0f417db38424)