Overhauls the security sandbox of Noodle
We now have a system wide setting if the sandbox is enabled at all.
Additionally one or more SandboxDetectors can determine if a template
is to be sandboxed or not.

Additionally, instead of just enabling or disabling the sandbox, we now
also support a "WARN_ONLY" mode, which enables the sandbox but
only reports warnings.

Finally the rules and the way we evaluate them (along with annotations)
have been made a bit more flexible (and some standard have been
added).