Prevents duplicate back navigation on page reload
Chrome has some weird behaviour where it completes running all code of the current execution context (everything that wasn't postponed by setTimeout or async calls).
This introduced a bug in the tycho history handling where a reload on a job error page (and similar scenarios) would actually jump all the way back to the previously open page (e.g. the jobs overview page).
We now use a simple flag to ensure the appendHistoryUrl logic does not get executed when we previously issued a back navigation in the on page load section of the history handling.

Fixes: SIRI-716