Only calls state handler restore when entry is actually changed
Optimized the 'handleStateChange' function to ensure that the state handler is only called when there is a real change in the state.
For this a deep comparison function 'deepEqual' was also added to make sure the changes in the nested properties of objects are correctly detected.
This should improve the "optical stability" of the page as state handlers aren't executing code that recreate the currently visible state (e.g. by executing the same search again).

Fixes: OX-10141