Correctly applies fallback logic for IE 11
Changed the fallback mechanism for silent form submission.
Originally, the functionality directly called `HTMLFormElement.submit` in cases when the `HTMLFormElement.requestSubmit` was not supported by the browser.
However, this bypassed triggering the 'onsubmit' event and constraint validation, and there was no way to prevent the page from reloading.
Now, a hidden submit button is created dynamically to handle the submission instead, ensuring 'onsubmit' event, validation, and prevention of page reload. This change enhances compatibility with older browsers like IE 11.

Fixes: OX-10197