Removes the X-Custom-TTL header
as proxy configuration is getting overly complex and demands "hacks" in order to alter the Date and Expires headers since we would deliver a file already expired as we cache longer than its age.

In fact, we see no need for a reverse-proxy to cache an asset longer than what we already set in its "Expire" and "Cache-Control" headers.

Fixes: SIRI-728