Do not reveal product in http.server header
Nowadays, it's unsafe to reveal which server is being used behind a web-site, such as Apache, nginx etc. (in our case we claim to use Netty) as attackers might profit from known security flaws in a specific product.

Therefore, we no longer advertise what we use in background. For debugging purposes, we do keep the host name delivering the content.