Always displays copy tooltip to the left 👈
By default, the tooltips are placed above the trigger element. However, the copy button has a high probability to be aligned to the right edge of the screen. The default caption "Copy value" is then too long to be displayed towards on the left. When clicking the button, the caption will be replaces by "Copied!" which is short enough to be displayed on top, leading the tooltip to jump.

In an intermediate solution, the initial tooltip was shortened to just "Copy". While this also solves the problem of the jumping tooltip, the difference to "Copied!" is very small and may be missed when not properly paying attention. Force-aligning to the left seems to be the better solution, as ist provides clearer feedback.

SIRI-924