Lazy usage of fallback element
We should only create the fallback when we really need it, otherwise we are using up unnecessary resources.

We do not need the fallback for images, as it is only needed anyway if the image cannot be loaded. However, the fallback would only try to load the image again. In this case, we can return NULL directly.

Fixes: SE-13398