Logs aggregatable route in page impression events
The previously logged string representation of the route matching pattern is hard to read and also not as easy to aggregate.
Now we actually log the route URI as it is defined in the Routed annotation of each route.

Its known and accepted that this means, that old and new events for the same route are not easily aggregatable together.

Fixes: OX-10830